sharkd: Add "comments" field to "frames" method
Update `sharkd`'s `frames` method to include a `comments` string array
containing the comments for a frame.  The existing boolean `ct` field,
which is `true` if the frame has one or more comments, has not been
modified.  Like the existing `ct` field, the new `comments` field is
only included in the response if the frame has comments.  A new
`test_sharkd_req_frames_comments` unit test has been added to verify
the functionality of the new `comments` field.