Qt: Make it possible to clear Comments from all interfaces
We only add an entry for an interface if it has a non-empty Comment
(aka user description). That means we need to make sure that the
map has a default empty QStringList for that pref so we can set it
to the empty string; otherwise, if some interfaces have Comments, but
the user tries to the clear the comments from all interfaces, it won't
work. The map won't have an entry for that pref, so it won't get
changed to the empty string.

The same is true for the Monitor Mode setting.