Qt: Make it more obvious and user-friendly to restore snaplen to default
In InterfaceTreeTelegate allow the editor QSpinBox for the snaplen column
to go from 0 to WTAP_MAX_PACKET_SIZE_STANDARD.
Set 0 to have the special value text of "default".

When the editor returns 0 or WTAP_MAX_PACKET_SIZE_STANDARD, display "default"
after viewing the results.

This allows entering "0" as an entry and it will be accepted as default.
Enterting WTAP_MAX_PACKET_SIZE_STANDARD (262144) will also work.
It will also display the string "default" for 0 when spinning the spinbox,
though not for 262144. (It's possible to subclass QSpinBox and override
textFromValue and valueFromText to do that, though then there are two
consecutive steps when wrapping around that both are "default.")

Fix #15389