btatt: Fix direction of Mesh Proxy Data In
Mesh Provisioning Data In and Mesh Proxy Data In are the
reverse direction of a duplex link, and are for the opposite
direction of the original request.

See the file in #15812 for an example. (Note that the retransmission
detection must be off for it to work, and the keys mentioned in the
issue must be configured.)

Fixup a4904bc1dd9f1866f04f2aaabbe164e73b59bce1