tap-iostat: Reduce memory usage
Each column only has a single calculation type and hf_index, so
we don't need to store that in each io_stat_item_t, but at the
per-column level. For the same reason, we can use a union for
the counter, since the statistic will only need one type.

This reduces the io_stat_item_t for the tshark tap from 80 B to
56 B on an LP64 platform.

Somewhat related to #10068, though it shouldn't have possibly
allocated that much memory anyway on a short file, which implies
that something odd was going on with macOS 10.9 Mavericks.