Qt: Update last applied filter if and only if applying
If we apply a display filter, and the filter is empty, we
want to store the empty string in last applied as well, so
we don't present placeholder text suggesting that a filter
is still being applied.

If we're only preparing a filter, then we don't want to
store that in the last_applied_ string, because then if someone
erases the filter without applying it we'll present placeholder
text suggesting that the filter is currently being applied.

Always update the last applied filter in applyDisplayFilter
if we're applying it (not if it's invalid and we return early.)