Qt: Adjust stretching for Edit Column
We want to use stretch factors, beccause the column definition
line edit for custom columns has larger needs than the column
title.

However, if those two elements have stretch factors but others
do not, then they will be allowed to shrink to their minimum before
the other widgets that don't have the stretch factor. We don't,
however, want the other widgets to stretch beyond their hinted
size, which in this case is a maximum.

So make the occurrenceLineEdit also have a stretch factor, but
set its maximum to its size hint. Do something similar with the
type combobox, except put a minimum size on it that is large
enough to view most types (with the date and time formats being
an exception.) For Windows (and possibly other platforms), force
the pop-up view to have the full width instead of eliding the text
as it does by default. (On many Linux desktops the pop-up always
has the full necessary width.)

The fieldsNameLineEdit is only relevant if the typeComboBox is
COL_CUSTOM. The text for "Custom" is small. So when
COL_CUSTOM is selected, shrink the size of the typeComboBox to what is
necessary for "Custom" and give extra space to the fieldsNameLineEdit.
For any other column type, do the reverse.

Fix #19458