Qt: Allow the Learn links to word wrap
Over the course of the 4.0 series the Sharkfest, Discord, and Donation
links were added to the Welcome page. They are now the limiting factor
in the application page width. For some language translations they
are even longer.

Some of the "Learn" links have multiple words, like "Questions and
Answers," or "Wireshark Discord." Allow the multiword link titles
to word wrap.

Fix #19861