capture: Support the -F option, request nanosecond time stamp for pcap
Support the -F option from the command line for dumpcap and wireshark,
only supporting pcapng and pcap.  Deprecate the -n and -P options.
Make the output resemble that from libwiretap, but implement it
separately because dumpcap intentionally does not link to libwiretap.

Request nanosecond time stamps for the pcap format as well; we believe
the vast majority of programs should be able to handle them at this
point. The Internet Draft of the pcap format specification treats them
as the same format:
https://datatracker.ietf.org/doc/html/draft-ietf-opsawg-pcap

For now libwiretap still treats them as different formats, which allows
converting explicitly between them; this can be used as a workaround if
someone needs to convert to the microsecond time stamp pcap format.
If someone requests nanosecond time resolution to tshark using the -F
flag, allow it.

Update the documentation.

Fix #18009