Qt: Fix a crash in Endpoints/Conversations if recentList is NULL
In strange situations the recent list might be NULL. There also
might be a situation where the protocol at the beginning of the
recent list doesn't exist by that short name anymore, if the user
has a recent file from a different version of Wireshark.

Also, don't both removing a tab if we'll automatically add it back
in. Perhaps this logic should be different not to automatically
add back in the tab, but at least it's better not to reset the
model only to leave it back as it was.

Fix #19869