DNS: Add new top-level "Answer Type" statistic
Add new top-level "Answer Type" statistic which counts the different
types of RR types (A, AAAA, PTR, etc.) in the answers section of the
DNS message.  Note the "Answer Type" is computed from all answers in
the DNS message, unlike the existing "Query Type" statistic which is
computed from just the first question.