merge: Don't abort a merge on a short read
If the error reading a capture file during a merge is
WTAP_ERR_SHORT_READ, that's a fatal error as far as reading from
_that_ file, but it doesn't mean we have to abort the entire merge.
It's a relatively normal condition from truncated files.

Report the error, stop processing that file, but keep going with the
merge.

Fix #15379 (at least the example shown there; we might want to
keep on trucking through WTAP_ERR_BAD_FILE too.)