capinfos: Fix printing packet comment headers with -TM
Commit 45772de9 added support for printing packet comments to
`capinfos` with new flags `-p`/`-P` as well as including them in the
machine-readable `-M` and table report machine-readable `-TM` outputs.
Unfortunately, the commit contained a bug when `-TM` was specified
causing none of the `Packet %d Comment` headers for packet comments
being printed.  This made the output decidedly less machine-readable.
This commit fixes this bug, ensuring that `-TM` output includes the
correct headers for captures containing packet comments.

Previously, in `print_stats_table_header` we used `g_malloc0` to
allocate a buffer `buf` large enough to hold the `Packet %d Comment`
header for any packet with a frame number from 1 to 2^64.
Unfortunately, since this buffer was initialized with `NULL` bytes,
using `strlen(buf)` as the `n` argument to `snprintf` when writing the
`Packet %d Comment` header to `buf` always led to it writing 0 bytes
to `buf`.  This is turn led `print_stats_table_header_label` to always
get an empty string as its argument, meaning no `Packet %d Comment`
header was printed.  The result of this bug was that the actual packet
comment values were printed, but without the accompanying `Packet %d
Comment` header.

Below is an example of the previous, broken behavior:

    $ capinfos -TM comment-test.pcapng
    File name	File type	File encapsulation	File time precision	Packet size limit	Packet size limit min (inferred)	Packet size limit max (inferred)	Number of packets	File size (bytes)	Data size (bytes)	Capture duration (seconds)	Start time	End time	Data byte rate (bytes/sec)	Data bit rate (bits/sec)	Average packet size (bytes)	Average packet rate (packets/sec)	SHA256	SHA1	Strict time order	Capture hardware	Capture oper-sys	Capture application	Capture comment
    /root/comment-test.pcapng	pcapng	ether	microseconds	(not set)	n/a	n/a	19	3944	2990	2.660439	2022-12-24 17:31:05.387079	2022-12-24 17:31:08.047518	1123.87	8991.00	157.37	7.14	e3ab39da3be87894a58e622dd08505b94f37bc5e478099037dcd03333a87c8b3	5afe04e6f4d5f485aad8b38ee42c54ee4fd6256e	True			Editcap (Wireshark) 3.6.19 (Git commit c38a01f797b1)		Frame two comment	Frame 6 has a comment	Frame 10 with _markdown_ **markdown**\r\n\r\n## Headers too\r\n\r\n[and a link!](https://www.google.com)	Frame 14 comment multiline\n\nHello there!

and here is an example of the new, fixed behavior:

    $ capinfos -TM comment-test.pcapng
    File name	File type	File encapsulation	File time precision	Packet size limit	Packet size limit min (inferred)	Packet size limit max (inferred)	Number of packets	File size (bytes)	Data size (bytes)	Capture duration (seconds)	Start time	End time	Data byte rate (bytes/sec)	Data bit rate (bits/sec)	Average packet size (bytes)	Average packet rate (packets/sec)	SHA256	SHA1	Strict time order	Capture hardware	Capture oper-sys	Capture application	Capture comment	Packet 2 Comment	Packet 6 Comment	Packet 10 Comment	Packet 14 Comment
    /home/niels/Downloads/comment-test.pcapng	pcapng	ether	microseconds	(not set)	n/a	n/a	19	3944	2990	2.660439	2022-12-24 12:31:05.387079	2022-12-24 12:31:08.047518	1123.87	8991.00	157.37	7.14	e3ab39da3be87894a58e622dd08505b94f37bc5e478099037dcd03333a87c8b3	5afe04e6f4d5f485aad8b38ee42c54ee4fd6256e	True			Editcap (Wireshark) 3.6.19 (Git commit c38a01f797b1)		Frame two comment	Frame 6 has a comment	Frame 10 with _markdown_ **markdown**\r\n\r\n## Headers too\r\n\r\n[and a link!](https://www.google.com)	Frame 14 comment multiline\n\nHello there!