log3gpp: Fix read of uninitialized data.
Clang static analysis reports that subsecond_decimals_buf[1] through [3] might
be used uninitialized.  This is because the array is being NUL-terminated, but
the NUL-termination is not relevant (and is in fact erroneous) when the array
is being directly evaluated one character at a time and assumed to be only
digits.

The NUL-termination can be removed (and the array size reduced by one) and
instead we directly initialize to '0' any characters not read from the source
line.