ui: Fix help url comments
We used to open HTML Help on Windows, in which case the functions
that retrieved a URL returned NULL. We don't do that anymore
(6bfab69d14d39863730f1c6c7bd278ec9f1e1673), which makes the text
about "If the attempt fails NULL will be returned" confusing.

In fact these functions are guaranteed to return a non null value,
even in the unexpected and defaults cases.