Remove some unnecessary extra layers of dissector.
Don't register dissectors that return TRUE or FALSE as as non-heuristic
dissectors; non-heuristic dissectors are supposed to return 0 if they
don't think the packet is for them and the number of bytes dissected if
they do.  Let the heuristic RTP dissector test whether those dissectors
return a zero or non-zero result.

(Ultimately, dissectors should be able to indicate "I dissected this"
vs. "this isn't mine" in some way that doesn't involve the number of
bytes dissected, as the number of bytes dissected can be 0, as per
issues #12366 and #12368.  That would:

* allow heuristic dissectors to return the number of bytes dissected, in
case a caller needs that;
* allow eliminating the whole distinction between heuristic and
non-heuristic dissectors, as they could have the same signature.)