Qt: Have all DisplayFilterEdits have bookmarks
Always add the bookmark button to a DisplayFilterEdit, such as
in the I/O Graphs dialog, the Coloring Rules dialog, etc.

Add a method to enable and disabled displaying bookmarks for,
e.g., the search frame which switches its line edit between
searching for display filters and other options (hex, string).

Make sure to update the filters if the app has been initialized
(we can't do this before it's been initialized, but DFEs from
other dialogs are created after initialization.)

Change the bookmark button placement from in the margin to the padding
of the box model, so that it works properly when contained in just
a DisplayFilterEdit and not a DisplayFilterCombo.

Only add the context menu item about the filter button preferences for
the main window display filter combobox to apply.

Fix #12028