Qt: Add filter buttons to QToolBar as QActions
QToolBar only adds items to the extension menu pop-up used when
the width is too small if they are added as QActions and QToolBar
created its own QToolButton for them. If we create our own
QToolButton and add it as a widget (or use a QWidgetAction), it
does not add them to the extension window.

Note that the default behavior is QToolButton::MenuButtonPopup
which does nothing if the button text is pressed, only if the
menu symbol (downwards pointing triangle) is pressed. When we
created our own QToolButton we made pressing the text also
open the menu. That seems much harder with the button created
by the tool bar internally. However, having buttons not appear
at all is worse.

Fix #19887.