tools: convert eti2wireshark.py
Convert eti2wireshark.py for the changes to hf and ett initialization.
Also convert to C99 types while we're here (#19116)

Update for a few of our other changes, like registering the dissector
handle to make it available for other dissectors.

Add the XML source from the current version used for the dissectors and
regenerate the dissectors, which is version 10. There are later
versions, currently up to 12. Fix the spelling in the XML in a few
places to match our own spelling corrections.

Add a basic README showing the script calls to regenerate the
dissectors.

This should be added to CMake.