Qt: fix unreadable syntax filter in dark mode on KDE with Qt 6
KDE Plasma 6.1.0 with the Breeze Dark theme on Qt 6.7.1 and Windows 11
with a high contrast theme have an unreadable display filter field. It
displays white text on a bright green or red background, see #18246.

This was caused by an invalid assumption that `colorScheme()` only
returns two values, Light and Black. It can also return Unknown which
does not imply Light. Make sure to fallback and check the actual colors.

Link: https://gitlab.com/wireshark/wireshark/-/issues/18246#note_1958637216
Link: https://code.qt.io/cgit/qt/qtbase.git/tree/src/gui/platform/unix/qgenericunixthemes.cpp?h=6.7.2#n1132
Link: https://code.qt.io/cgit/qt/qtbase.git/tree/src/plugins/platforms/windows/qwindowstheme.cpp?h=6.7.2#n549
Fixes: v4.1.1rc0-163-g0573bbaed6 ("Qt: Dark mode updates")