WiSUN: correct offset logic errors
Change local variable `offset` from a `guint8` to a `guint` in two
functions with loops, to prevent a potential infinite loop from
overflowing the variable.

Change `dissect_wisun_nrie()` to return the number of bytes it
dissected, since that's how `dissect_wisun_hie()` seemingly expects it
to behave. It previously returned the new offset at the end of what it
dissected -- which is how several `dissect_...()` functions in this
dissector are expected to behave, but this is not one of them. (This
inconsistency adds friction when maintaining this dissector and should
be addressed at some point.)