radius: Update dictionaries to FreeRADIUS 3.2.4
These are the easy cases - entirely new vendor dictionaries, added
attributes, or in some cases simply updating the copyright, license,
and hash from FreeRADIUS. There's also a few dictionaries that have
been previously added to the Wireshark repo but not added to the default
dictionary config (even though they work fine), as well as some
reordering of the dictionaries to match the lexicographical ordering
that FreeRADIUS uses.

Also update the internal FreeRADIUS dictionary and the DHCP dictionary
(neither of which we actually use nor support everything in them), just
to stay consistent with FreeRADIUS.

There's a number of dictionaries where there are Wireshark-specific changes,
or changes taken from FreeRADIUS master that are not in FreeRADIUS stable
(3.2.4). Those are not in this commit.

Add CONFIGURE_DEPENDS to the radius GLOB in CMake so that it will pick
up new files. That requires CMake 3.12 (3.13 is our minimum.)
https://cmake.org/cmake/help/latest/command/file.html#glob

Ping #15458