radius: Support sub-TLVs of sub-TLVs
Support sub-TLVs inside sub-TLVs inside sub-TLVs, up to a nesting limit
of 10 (not even the WiMAX dictionary goes that deep.)

Update the WiMAX dictionary, which needs this since there are several
nested levels of TLVs.

For now we still use our alteration so that the parser can handle it.
It would be nice to be able to parse the FreeRADIUS format without
changes for TLVs.

Fix #7050. Ping #15458.