Add G_GNUC_ALLOC_SIZE annotations to wmem functions.
This Glib macro uses __attribute__((alloc_size(n))) to aid static analysis of
buffer bounds.  Using it on wmem_alloc, wmem_alloc0, wmem_realloc and
wmem_memdup gives the compiler's static analyzer a chance to detect
out-of-bounds accesses to the buffers they return.

It would be nice to extend this to some of the tvb functions, but the tvb API's
use of length -1 to indicate "to the end of the tvbuff" gets in the way.

This annotation surfaced the issue addressed in commit 54410038.