merge: squelch some compiler warnings.
The number of files to be processed by merge_files_common() should never
be zero, so the loop over all files should be executed at least once.

merge_open_in_files() is called for each file; if it returns 0, it sets
err_fileno, so that shouldn't result in err_fileno being used when not
set.

merge_process_packets() is also called, and if it returns anything other
than MERGE_OK, it will set err, err_info, err_fileno, and err_framenum,
so it's not clear why, on Linux Mint, the compiler thinks err_framenum
is being used without being set.

Perhaps the assertion test for in_file_count > 0 isn't being done
because NDEBUG is set, and it thinks you might never execute the loop;
however, in that case, status, which is initialized to MERGE_OK, will
still be MERGE_OK, so the error-reporting code should never be executed.

For now, we just initialize err, err_fileno, and err_framenum to 0 and
err_info to NULL, given that we're already initializing status to
MERGE_OK.

While we're at it, remove some unnecessary settings of err_framenum.