WSDG: update help messages.
Update the capinfos page to reflect changes to correctly describe what
-a and -e do.

Update the dumpcap page not to mention rpcap, as it's supported only if
the pcap library supports it, and the only platform on which it's
supported by default is Windows.

Update the dumpcap and TShark man pages to remove a Linux-specific
message.