pidl: Convert generated types to C99
There are a few types, such as pointers, that have different wire
lengths in NDR and NDR64. In NDR64 they are padded for alignment
for performance reasons, even though the amount of useful data is
the same.

mapScalarType in Parse::Pidl::Typelist returns the size of the
useful data. The Wireshark NDR dissectors retrieve the entire
value on the wire (the padding should be null, though for fuzzed
data it isn't), so we need a map converting to the C type on the
wire including padding.

Use that and regenerate the dissectors to map to C99 types instead
of glib types (and our typedef guint1632 / guint3264.) Some of
the non generated dissectors still use that typedef.

Ping #19116