editcap: Fix handling of unrecognized long options
The common ws_getopt_long handles reporting if options are
unrecognized or require arguments, so we don't need to do
it again in editcap, so we can remove the leading colon in
the optstring. This allows ws_getopt to handle it better
when an unrecognized long option is used.

Currently it prints out a null by trying to output the single-letter
option for unrecognized long options, i.e.

    editcap: invalid option -- ''

instead of the long option actually input.