editcap: Add extract-secrets
Add an extract-secrets option to editcap, to write the information from
DSBs back to file(s). It takes over the output file and ignores all
other options. If there are multiple DSBs, each is written to a
separate file (some types could be concatenated possibly, but not all.)

Document and add a test.

Fix #18197