Qt: DisplayFilterEdit bookmark button background
Make the DisplayFilterEdit have the same border (and border color)
as the vertical dividing line between the bookmark button and the
filter, just as the DisplayFilterCombo does.

Add a tiny margin to the button style, so that the 1px border from
the DisplayFilterEdit shows up even if the button background is
not transparent. Make the button background be that of the line
edit, so that it doesn't get overriden with the syntax colors.

Followup to !16002